images/solution/requirements.txt
mysqlclient ~= 2.1.0
SQLAlchemy ~= 1.4.31
python-dotenv~=0.21.1
pandas~=1.5.2